+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+ - Glium now automatically calls `glDraw*BaseVertex` if it is supported.
  - Added the `CapabilitiesSource` trait implemented automatically on all types that implement `Facade`.
  - Added `is_supported` to `IndexType`, `Index`, `PrimitiveType`, `Attribute` and `AttributeType`.
  - The `implement_buffer_content!` and `implement_block_layout!` macros can now take a struct with a lifetime parameter by passing `Foo<'a>` instead of `Foo`.